begin evasion detector refactoring
