+++ art/defences/detector/evasion/__init__.py
@@ -1,6 +1,5 @@
 """
 Module implementing detector-based defences against evasion attacks.
 """
-from art.defences.detector.evasion import subsetscanning
-
-from art.defences.detector.evasion.detector import BinaryInputDetector, BinaryActivationDetector
+from art.defences.detector.evasion.binary_detector import BinaryInputDetector, BinaryActivationDetector
+from art.defences.detector.evasion.subsetscanning import SubsetScanningDetector
