--- docs/modules/defences/detector_evasion_subsetscanning.rst
@@ -1,8 +1,0 @@
-:mod:`art.defences.detector.evasion.subsetscanning`
-===================================================
-.. automodule:: art.defences.detector.evasion.subsetscanning
-
-Subset Scanning Detector
-------------------------
-.. autoclass:: SubsetScanningDetector
-   :members:
